+++ app/controllers/api/api_controller.rb
@@ -2,6 +2,7 @@
 
 module Api
   class ApiController < ActionController::API
+    include VerifyTokens
     def t(key, options = {})
       I18n.t("#{controller_name}_controller.#{key}", **options)
     end
