+++ app/controllers/api/users_controller.rb
@@ -1,7 +1,8 @@
 # frozen_string_literal: true
 
 module Api
-  class UsersController < AuthenticateController
+  class UsersController < ApiController
+    before_action :require_access_token
     before_action :validate_user_id, only: %i[show]
     # GET /api/users/:id
     def show
